IDのキャストミスの線を当たっているが、直らない

+++ app/src/main/java/com/timeshappy/scoresheet/Item.java
@@ -1,9 +1,6 @@
 package com.timeshappy.scoresheet;
-
-
-
 class Item {
-    private int id;
+    private long id;
     private String count_timer;
     private long date;
     private String number;
@@ -18,12 +15,7 @@ class Item {
 
     public static String numberT;
 
-
-
-
-
-
- public Item(int id, String count_timer, long date, String number, int foul, int teamA_foul, int teamB_foul,String teama_name,String teamb_name, int goal_a, int goal_b,String foul_name) {
+ public Item(long id, String count_timer, long date, String number, int foul, int teamA_foul, int teamB_foul,String teama_name,String teamb_name, int goal_a, int goal_b,String foul_name) {
         this.id = id;
         this.count_timer = count_timer;
         this.date = date;
@@ -37,15 +29,11 @@ class Item {
         this.goal_b = goal_b;
         this.foul_name = foul_name;
     }
-
-
-
-
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
